Change type of Quantity in db. Change method from getInt to getFloat in DatabaseAccess. Remove button Show Meal in activity_main. Add button Add to List in activity_show_meal.

+++ app/src/main/java/com/example/noshapp/DatabaseAccess.java
@@ -72,7 +72,7 @@ public class DatabaseAccess {
         c.moveToFirst();
         while (!c.isAfterLast()) {
             if (c.getString(c.getColumnIndex("Name")) != null) {
-                Ingredients ingredient = new Ingredients(c.getString(c.getColumnIndex("Name")), c.getInt(c.getColumnIndex("Quantity")), c.getString(c.getColumnIndex("Unit")), c.getInt(c.getColumnIndex("Meal_ID")));
+                Ingredients ingredient = new Ingredients(c.getString(c.getColumnIndex("Name")), c.getFloat(c.getColumnIndex("Quantity")), c.getString(c.getColumnIndex("Unit")), c.getInt(c.getColumnIndex("Meal_ID")));
                 ingredients.add(ingredient);
             }
             c.moveToNext();
